app_service/requirements.txt
# FastAPI and Uvicorn
fastapi==0.109.2
uvicorn==0.27.1

# Database Connector for PostgreSQL
asyncpg==0.29.0

# For loading .env files
python-dotenv==1.0.1

# Twilio (if still needed for other parts of your project, otherwise remove)
twilio

# Flask (if still needed for other parts of your project, otherwise remove)
flask

# For form data, often used with FastAPI file uploads
python-multipart==0.0.9

google-generativeai==0.3.2

boto3
google-cloud-aiplatform

pydantic==2.6.1
email-validator==2.1.0.post1

pytz==2024.1
